Implimented rank()
also changed the way size() works. Now every node stores the size of
it's own subtree, and the size of the tree is the size of the root.
Also implemented some helping mehtods and fixed some random annoying
issues like names of methods, variables, etc. as well as some leftover
bugs.